Fix tests

+++ engine/src/main/kotlin/world/gregs/voidps/engine/entity/character/player/skill/Skills.kt
@@ -30,7 +30,7 @@ interface Skills {
         experience.add(handler)
     }
 
-    fun blockedExperience(handler: Player.(skill: Skill, exp: Double) -> Unit) {
+    fun blockedExperience(handler: Player.(skill: Skill, exp: Int) -> Unit) {
         blockedExperience = handler
     }
 
@@ -41,7 +41,7 @@ interface Skills {
         private val playerMaxChanged = Object2ObjectOpenHashMap<Skill?, MutableList<(Player, Skill, from: Int, to: Int) -> Unit>>(15)
 
         private val experience = ObjectArrayList<(Player, Skill, Int, Int) -> Unit>(5)
-        private var blockedExperience: (Player.(Skill, Double) -> Unit)? = null
+        private var blockedExperience: (Player.(Skill, Int) -> Unit)? = null
 
         fun exp(player: Player, skill: Skill, from: Int, to: Int) {
             for (handler in experience) {
@@ -49,7 +49,7 @@ interface Skills {
             }
         }
 
-        fun blocked(player: Player, skill: Skill, exp: Double) {
+        fun blocked(player: Player, skill: Skill, exp: Int) {
             blockedExperience?.invoke(player, skill, exp)
         }
